split test_torch
